updater: Force dracut run
Ignore failure to update portage and proceed to try updating world. Add
the `force_kernel_install` fact to the kernel build run to force
initramfs creation to pull in, for example, updated ZFS modules, even if
the kernel sources, controlled by the `build=kernel` fact, are not
updated. Bump version to 0.5.9.

Closes #24